Commit of sync-merged devDanFit/standalone.
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSTANDALONE/branches/devDanFit@789 9ff7d5ee-7a19-0410-ab54-ec97d5369f81

+++ GEMS3K/ipm_main.cpp
@@ -31,6 +31,7 @@
 
 #include<iomanip>
 
+#include "node.h"
 #include "m_param.h"
 #include "jama_lu.h"
 #include "jama_cholesky.h"
@@ -870,8 +871,8 @@ STEP_POINT("FIA Iteration");
    {  // Strict mode of mass balance control
        iRet = 2;
        char buf[320];
-       sprintf( buf, "(MBR(%ld)) Maximum allowed number of MBR iterations (%d) exceeded! ",
-                WhereCalledFrom, pa->p.DP );
+       sprintf( buf, "(MBR(%ld)) Maximum allowed number of MBR iterations (%ld) exceeded! ",
+                WhereCalledFrom, (long int)pa->p.DP );
        setErrorMessage( 4, "E04IPM: Mass Balance Refinement: ", buf );
        return iRet; // no MBR() solution
    }
